mlops1/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
fsspec==2025.2.0
gradio==5.17.1
gradio_client==1.7.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
mss==10.0.0
numpy==2.2.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.50.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python3-xlib==0.15
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rich==13.9.4
ruff==0.9.7
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.1
starlette==0.45.3
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
wheel==0.45.1
zmg==0.1
zmq==0.0.0
